% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.2 analogous.R
\name{analogous_palette}
\alias{analogous_palette}
\title{Generate an Analogous Color Palette}
\usage{
analogous_palette(
  base_color = NULL,
  spread = 15,
  n = 6,
  ctc_colors = TRUE,
  show_pal = TRUE
)
}
\arguments{
\item{base_color}{A character string specifying the base hexadecimal color code
(e.g., "#ECD452"). If \code{NULL}, a random color is selected from the \code{chinacolor}
dataset.}

\item{spread}{A numeric value specifying the angular spread of the hues on the
color wheel in degrees. Defaults to 10.}

\item{n}{An integer specifying the number of colors in the palette.
It is highly recommended to be between 3 and 7. Defaults to 6.}

\item{ctc_colors}{A logical value. If \code{TRUE}, the generated colors are mapped
to the closest  "Chinese traditional colors" from the \code{chinacolor} dataset.}

\item{show_pal}{A logical value. If \code{TRUE}, the generated palette will be
plotted.}
}
\value{
A character vector of hexadecimal color codes with the following attributes:
\itemize{
\item \code{type}: Character string indicating the palette type, always "analogous"
\item \code{n}: Number of colors in the palette
\item \code{ctc_colors}: Logical indicating if Chinese traditional color correction was applied
}
}
\description{
This function creates an analogous color palette based on a specified
or randomly selected base color.
}
\examples{
 \dontrun{
# Generate a palette based on a specific color
analogous_palette (base_color = "#779649",
         spread = 30,
         n = 7,
         show_pal = TRUE,
         ctc_colors = T)
# Generate a palette using a random base color
analogous_palette (n = 6,
         spread = 35,
         show_pal = TRUE,
         ctc_colors = T)
}
}
