% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_l_val}
\alias{get_l_val}
\title{Internal: Generate lightness values for color palette}
\usage{
get_l_val(n = NULL, base_lightness = NULL, lightness_range = c(15, 95))
}
\arguments{
\item{n}{Number of lightness values to generate}

\item{base_lightness}{The central lightness value to build the sequence around}

\item{lightness_range}{Numeric vector of length 2 specifying min and max lightness (0-100)}
}
\value{
A numeric vector of lightness values
}
\description{
This internal function calculates a sequence of lightness values centered around
a base lightness value within the specified range. It ensures the generated
values are evenly distributed while prioritizing the base lightness position.
}
\keyword{internal}
