% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/7.1 monochromatic.R
\name{monochromatic_palette}
\alias{monochromatic_palette}
\title{Generate a monochromatic color palette}
\usage{
monochromatic_palette(
  base_color = NULL,
  lightness_range = c(25, 85),
  n = 6,
  ctc_colors = FALSE,
  show_pal = TRUE
)
}
\arguments{
\item{base_color}{A character string specifying the base hexadecimal color code
(e.g., "#ECD452"). If \code{NULL}, a random color is selected from the \code{chinacolor}
dataset.}

\item{lightness_range}{A numeric vector of length 2 specifying the desired
minimum and maximum lightness values (L in HCL, 0-100) for the palette.
Defaults to \code{c(25, 85)}.}

\item{n}{An integer specifying the number of colors in the palette.
Defaults to 6.}

\item{ctc_colors}{A logical value. If \code{TRUE}, the generated colors are mapped
to the closest  "Chinese traditional colors" from the \code{chinacolor} dataset.}

\item{show_pal}{A logical value. If \code{TRUE}, the generated palette will be
plotted.}
}
\value{
A character vector of hexadecimal color codes with the following attributes:
\itemize{
\item \code{type}: Character string indicating the palette type, always "sequential"
\item \code{n}: Number of colors in the palette
\item \code{ctc_colors}: Logical indicating if Chinese traditional color correction was applied
}
}
\description{
This function creates a continuous, monochromatic color palette
based on a specified or randomly selected base color.
}
\details{
The function generates a palette by varying the lightness of a base color
while maintaining its hue and relative chroma. It leverages the HCL color
space for perceptual uniformity. The palette can be directly displayed and
optionally "corrected to color-true" (\code{ctc}) by finding the closest match
in a predefined set of colors, such as those from Chinese color standards.
}
\examples{
\dontrun{
# Generate a palette based on a specific color
monochromatic_palette(base_color = "#ECD452",
             n = 6,
              lightness_range = c(30, 85),
              show_pal = TRUE)

# Generate a palette using a random base color
monochromatic_palette(n = 6,
              lightness_range = c(30, 95),
              show_pal = TRUE)
# palette without colors in `chinacolor` dataset.
monochromatic_palette(n = 6,
              lightness_range = c(30, 95),
              ctc_colors = F,
              show_pal = TRUE)
}
}
