% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{ordered_color_set}
\alias{ordered_color_set}
\title{Internal: Order color set by lightness}
\usage{
ordered_color_set(color_set, direction = 1)
}
\arguments{
\item{color_set}{Character vector of hexadecimal color codes}

\item{direction}{Sorting direction: 1 for light to dark, 2 for dark to light}
}
\value{
The input color set reordered by lightness
}
\description{
This internal function sorts a set of colors by their lightness value in
either ascending or descending order.
}
\keyword{internal}
