% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/8.scale_fill_color_ctc_c.R
\name{scale_fill_ctc_c}
\alias{scale_fill_ctc_c}
\title{Continuous Fill Scale for ggplot2 Using Built-in CTC Palettes or Custom Palette Objects (Outline Colors)}
\usage{
scale_fill_ctc_c(palette_name, direction = 1, ...)
}
\arguments{
\item{palette_name}{For built-in palette, the input as below are available:
\enumerate{
\item the name of palette(the element name for a palette in the palette_list),such as "seq01","div18",...;
\item index of the palette, 01-60;
\item the palette_name value in a palette, such as "earth_autumn",for Chinese name;
\item the palette_name_e value in a palette, such as "violet_bloom",for English name;
\item a custom palette object with attributes (e.g., from \code{ctc_palette()},\code{diverging_palette()},\code{monochromatic_palette()} or \code{analogous_palette()})
You can get the built-in palette_name values and a quick preview the palettes by run \code{list_palettes()} function.
}}

\item{direction}{Numeric (1 or -1) to control color order: 1 for original order, -1 for reversed. Defaults to 1.}

\item{...}{Additional arguments passed to \code{ggplot2::scale_color_gradientn()} (e.g., \code{name} for legend title).}
}
\value{
A ggplot2 continuous outline fill scale object.
}
\description{
Creates a continuous outline fill scale for ggplot2 using sequential or diverging palettes
from built-in CTC palettes (seq01-seq20, div01-div20, qual01-qual20) or custom palette  via \code{ctc_palette()},
or custom palette objects with "sequential" or "diverging" type attributes via \code{monochromatic_palette()} or \code{analogous_palette()},
Designed for mapping continuous data with smooth gradient interpolation.
}
\details{
\itemize{
\item Supports both built-in palettes (requires \code{palette_list} loaded via \code{data(palette_list)} first)
\item Custom palette objects with attributes.
\item Valid built-in palette names: sequential (seq01-seq20), diverging (div01-div20), qualitative (qual01-qual20).
\item Custom palette objects should have a "type" attribute ("sequential", or "diverging").
\item Warns if using qualitative palettes (designed for categorical data) with continuous data.
}
}
\examples{
\dontrun{
data(palette_list)  # Load built-in palettes

# 1 Use diverging palette for fill,
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
  geom_raster() +
  scale_fill_ctc_c(palette_name = "div12", direction = -1, name = "Density")

  # 2 palette by index value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
geom_raster() +
    scale_fill_ctc_c(palette_name = 33, direction = 1, name = "Density")

  # 3 palette by Chinese palette_name value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = "海天沙影", direction = 1, name = "Density")

  # 4 palette by English palette_name_e value
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = "ocean_rose", direction = -1, name = "Density")

  # 5 Using custom palette with attributes from diverging_palette()
custom_div_pal <- diverging_palette(base_color = "#336B6B",
                                    n = 8,
                                    ctc_colors = TRUE)

ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = custom_div_pal, direction = -1, name = "Density")
custom_pick <- create_color_pick(
    groups = c(12,13,76),
    subgroups = list(-1,1,c(1,2,3,4)),  # -1 reverses subgroup order (4:3:2:1)
    order_rule = 1)
custom_div_pal2 <- ctc_palette(
    type = "custom",
    color_pick = custom_pick,
    show_colors = TRUE,
    palette_type = "diverging",
    palette_title = "diverging palette"
)
ggplot(faithfuld, aes(x = eruptions, y = waiting, fill = density)) +
    geom_raster() +
    scale_fill_ctc_c(palette_name = custom_div_pal2, direction = 1, name = "Density")

}

}
\seealso{
\code{ctc_palette()},\code{diverging_palette()},\code{monochromatic_palette()} or \code{analogous_palette()}for palette generation;
\code{scale_colour_ctc_c()} for fill scales;
\code{plot_palettes()} to visualize valid palettes.
\code{list_palettes()} to display a searchable table of all available color palettes with their identifiers with color preview.
}
